[Decomp][Command] Cohesion split: extract player state + stat-modify commands to Player*Commands.cpp
