+++ src/game/ChatCommands/PlayerCommands.cpp
@@ -676,2097 +676,39 @@ void ChatHandler::HandleCharacterLevel(Player* player, ObjectGuid player_guid, u
     }
 }
 
-/**
- * @brief Handler for HandleReviveCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleReviveCommand(char* args)
-{
-    Player* target;
-    ObjectGuid target_guid;
-    if (!ExtractPlayerTarget(&args, &target, &target_guid))
-    {
-        return false;
-    }
 
-    if (target)
-    {
-        target->ResurrectPlayer(0.5f);
-        target->SpawnCorpseBones();
-    }
-    else // will resurrected at login without corpse
-    {
-        sObjectAccessor.ConvertCorpseForPlayer(target_guid);
-    }
 
-    return true;
-}
 
-/**
- * @brief Handler for HandleDismountCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleDismountCommand(char* /*args*/)
-{
-    Player* player = m_session->GetPlayer();
 
-    // If player is not mounted, so go out :)
-    if (!player->IsMounted())
-    {
-        SendSysMessage(LANG_CHAR_NON_MOUNTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
 
-    if (player->IsTaxiFlying())
-    {
-        SendSysMessage(LANG_YOU_IN_FLIGHT);
-        SetSentErrorMessage(true);
-        return false;
-    }
 
-    player->Unmount();
-    player->RemoveSpellsCausingAura(SPELL_AURA_MOUNTED);
-    return true;
-}
 
-/**
- * @brief Handler for HandleLinkGraveCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleLinkGraveCommand(char* args)
-{
-    uint32 g_id;
-    if (!ExtractUInt32(&args, g_id))
-    {
-        return false;
-    }
 
-    char* teamStr = ExtractLiteralArg(&args);
 
-    Team g_team;
-    if (!teamStr)
-    {
-        g_team = TEAM_BOTH_ALLOWED;
-    }
-    else if (strncmp(teamStr, "horde", strlen(teamStr)) == 0)
-    {
-        g_team = HORDE;
-    }
-    else if (strncmp(teamStr, "alliance", strlen(teamStr)) == 0)
-    {
-        g_team = ALLIANCE;
-    }
-    else
-    {
-        return false;
-    }
 
-    WorldSafeLocsEntry const* graveyard = sWorldSafeLocsStore.LookupEntry(g_id);
-    if (!graveyard)
-    {
-        PSendSysMessage(LANG_COMMAND_GRAVEYARDNOEXIST, g_id);
-        SetSentErrorMessage(true);
-        return false;
-    }
 
-    Player* player = m_session->GetPlayer();
 
-    uint32 zoneId = player->GetZoneId();
 
-    AreaTableEntry const* areaEntry = GetAreaEntryByAreaID(zoneId);
-    if (!areaEntry || areaEntry->zone != 0)
-    {
-        PSendSysMessage(LANG_COMMAND_GRAVEYARDWRONGZONE, g_id, zoneId);
-        SetSentErrorMessage(true);
-        return false;
-    }
 
-    if (sObjectMgr.AddGraveYardLink(g_id, zoneId, g_team))
-    {
-        PSendSysMessage(LANG_COMMAND_GRAVEYARDLINKED, g_id, zoneId);
-    }
-    else
-    {
-        PSendSysMessage(LANG_COMMAND_GRAVEYARDALRLINKED, g_id, zoneId);
-    }
 
-    return true;
-}
 
-/**
- * @brief Handler for HandleItemMoveCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleItemMoveCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-    uint8 srcslot, dstslot;
 
-    char* pParam1 = strtok(args, " ");
-    if (!pParam1)
-    {
-        return false;
-    }
 
-    char* pParam2 = strtok(NULL, " ");
-    if (!pParam2)
-    {
-        return false;
-    }
 
-    srcslot = (uint8)atoi(pParam1);
-    dstslot = (uint8)atoi(pParam2);
 
-    if (srcslot == dstslot)
-    {
-        return true;
-    }
 
-    Player* player = m_session->GetPlayer();
-    if (!player->IsValidPos(INVENTORY_SLOT_BAG_0, srcslot, true))
-    {
-        return false;
-    }
 
-    // can be autostore pos
-    if (!m_session->GetPlayer()->IsValidPos(INVENTORY_SLOT_BAG_0, dstslot, false))
-    {
-        return false;
-    }
 
-    uint16 src = ((INVENTORY_SLOT_BAG_0 << 8) | srcslot);
-    uint16 dst = ((INVENTORY_SLOT_BAG_0 << 8) | dstslot);
 
-    player->SwapItem(src, dst);
 
-    return true;
-}
 
-/**
- * @brief Handler for HandleCooldownCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleCooldownCommand(char* args)
-{
-    Player* target = getSelectedPlayer();
-    if (!target)
-    {
-        SendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
 
-    std::string tNameLink = GetNameLink(target);
 
-    if (!*args)
-    {
-        target->RemoveAllSpellCooldown();
-        PSendSysMessage(LANG_REMOVEALL_COOLDOWN, tNameLink.c_str());
-    }
-    else
-    {
-        // number or [name] Shift-click form |color|Hspell:spell_id|h[name]|h|r or Htalent form
-        uint32 spell_id = ExtractSpellIdFromLink(&args);
-        if (!spell_id)
-        {
-            return false;
-        }
 
-        if (!sSpellStore.LookupEntry(spell_id))
-        {
-            PSendSysMessage(LANG_UNKNOWN_SPELL, target == m_session->GetPlayer() ? GetMangosString(LANG_YOU) : tNameLink.c_str());
-            SetSentErrorMessage(true);
-            return false;
-        }
 
-        target->RemoveSpellCooldown(spell_id, true);
-        PSendSysMessage(LANG_REMOVE_COOLDOWN, spell_id, target == m_session->GetPlayer() ? GetMangosString(LANG_YOU) : tNameLink.c_str());
-    }
-    return true;
-}
 
-/**
- * @brief Handler for HandleSaveCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleSaveCommand(char* /*args*/)
-{
-    Player* player = m_session->GetPlayer();
 
-    // save GM account without delay and output message (testing, etc)
-    if (GetAccessLevel() > SEC_PLAYER)
-    {
-        player->SaveToDB();
-        SendSysMessage(LANG_PLAYER_SAVED);
-        return true;
-    }
 
-    // save or plan save after 20 sec (logout delay) if current next save time more this value and _not_ output any messages to prevent cheat planning
-    uint32 save_interval = sWorld.getConfig(CONFIG_UINT32_INTERVAL_SAVE);
-    if (save_interval == 0 || (save_interval > 20 * IN_MILLISECONDS && player->GetSaveTimer() <= save_interval - 20 * IN_MILLISECONDS))
-    {
-        player->SaveToDB();
-    }
 
-    return true;
-}
 
-/**
- * @brief Handler for HandleSaveAllCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleSaveAllCommand(char* /*args*/)
-{
-    sObjectAccessor.SaveAllPlayers();
-    SendSysMessage(LANG_PLAYERS_SAVED);
-    return true;
-}
 
-/**
- * @brief Handler for HandleStartCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleStartCommand(char* /*args*/)
-{
-    Player* chr = m_session->GetPlayer();
-
-    if (chr->IsTaxiFlying())
-    {
-        SendSysMessage(LANG_YOU_IN_FLIGHT);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (chr->IsInCombat())
-    {
-        SendSysMessage(LANG_YOU_IN_COMBAT);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // cast spell Stuck
-    chr->CastSpell(chr, 7355, false);
-    return true;
-}
-
-/**
- * @brief Handler for HandleTaxiCheatCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleTaxiCheatCommand(char* args)
-{
-    bool value;
-    if (!ExtractOnOff(&args, value))
-    {
-        SendSysMessage(LANG_USE_BOL);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (!chr)
-    {
-        chr = m_session->GetPlayer();
-    }
-    // check online security
-    else if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    if (value)
-    {
-        chr->SetTaxiCheater(true);
-        PSendSysMessage(LANG_YOU_GIVE_TAXIS, GetNameLink(chr).c_str());
-        if (needReportToTarget(chr))
-        {
-            ChatHandler(chr).PSendSysMessage(LANG_YOURS_TAXIS_ADDED, GetNameLink().c_str());
-        }
-    }
-    else
-    {
-        chr->SetTaxiCheater(false);
-        PSendSysMessage(LANG_YOU_REMOVE_TAXIS, GetNameLink(chr).c_str());
-        if (needReportToTarget(chr))
-        {
-            ChatHandler(chr).PSendSysMessage(LANG_YOURS_TAXIS_REMOVED, GetNameLink().c_str());
-        }
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleExploreCheatCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleExploreCheatCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int flag = atoi(args);
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (flag != 0)
-    {
-        PSendSysMessage(LANG_YOU_SET_EXPLORE_ALL, GetNameLink(chr).c_str());
-        if (needReportToTarget(chr))
-        {
-            ChatHandler(chr).PSendSysMessage(LANG_YOURS_EXPLORE_SET_ALL, GetNameLink().c_str());
-        }
-    }
-    else
-    {
-        PSendSysMessage(LANG_YOU_SET_EXPLORE_NOTHING, GetNameLink(chr).c_str());
-        if (needReportToTarget(chr))
-        {
-            ChatHandler(chr).PSendSysMessage(LANG_YOURS_EXPLORE_SET_NOTHING, GetNameLink().c_str());
-        }
-    }
-
-    for (uint8 i = 0; i < PLAYER_EXPLORED_ZONES_SIZE; ++i)
-    {
-        if (flag != 0)
-        {
-            m_session->GetPlayer()->SetFlag(PLAYER_EXPLORED_ZONES_1 + i, 0xFFFFFFFF);
-        }
-        else
-        {
-            m_session->GetPlayer()->SetFlag(PLAYER_EXPLORED_ZONES_1 + i, 0);
-        }
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleLevelUpCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleLevelUpCommand(char* args)
-{
-    int32 addlevel = 1;
-    char* nameStr = NULL;
-
-    if (*args)
-    {
-        nameStr = ExtractOptNotLastArg(&args);
-
-        // exception opt second arg: .levelup $name
-        if (!ExtractInt32(&args, addlevel))
-        {
-            if (!nameStr)
-            {
-                nameStr = ExtractArg(&args);
-            }
-            else
-            {
-                return false;
-            }
-        }
-    }
-
-    Player* target;
-    ObjectGuid target_guid;
-    std::string target_name;
-    if (!ExtractPlayerTarget(&nameStr, &target, &target_guid, &target_name))
-    {
-        return false;
-    }
-
-    int32 oldlevel = target ? target->getLevel() : Player::GetLevelFromDB(target_guid);
-    int32 newlevel = oldlevel + addlevel;
-
-    if (newlevel < 1)
-    {
-        newlevel = 1;
-    }
-
-    if (newlevel > STRONG_MAX_LEVEL)                        // hardcoded maximum level
-    {
-        newlevel = STRONG_MAX_LEVEL;
-    }
-
-    HandleCharacterLevel(target, target_guid, oldlevel, newlevel);
-
-    if (!m_session || m_session->GetPlayer() != target)     // including chr==NULL
-    {
-        std::string nameLink = playerLink(target_name);
-        PSendSysMessage(LANG_YOU_CHANGE_LVL, nameLink.c_str(), newlevel);
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleShowAreaCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleShowAreaCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    int area = GetAreaFlagByAreaID(atoi(args));
-    int offset = area / 32;
-    uint32 val = (uint32)(1 << (area % 32));
-
-    if (area < 0 || offset >= PLAYER_EXPLORED_ZONES_SIZE)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    uint32 currFields = chr->GetUInt32Value(PLAYER_EXPLORED_ZONES_1 + offset);
-    chr->SetUInt32Value(PLAYER_EXPLORED_ZONES_1 + offset, (uint32)(currFields | val));
-
-    SendSysMessage(LANG_EXPLORE_AREA);
-    return true;
-}
-
-/**
- * @brief Handler for HandleHideAreaCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleHideAreaCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    int area = GetAreaFlagByAreaID(atoi(args));
-    int offset = area / 32;
-    uint32 val = (uint32)(1 << (area % 32));
-
-    if (area < 0 || offset >= PLAYER_EXPLORED_ZONES_SIZE)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    uint32 currFields = chr->GetUInt32Value(PLAYER_EXPLORED_ZONES_1 + offset);
-    chr->SetUInt32Value(PLAYER_EXPLORED_ZONES_1 + offset, (uint32)(currFields ^ val));
-
-    SendSysMessage(LANG_UNEXPLORE_AREA);
-    return true;
-}
-
-/**
- * @brief Handler for HandleAddItemCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleAddItemCommand(char* args)
-{
-    char* cId = ExtractKeyFromLink(&args, "Hitem");
-    if (!cId)
-    {
-        return false;
-    }
-
-    uint32 itemId = 0;
-    if (!ExtractUInt32(&cId, itemId))                       // [name] manual form
-    {
-        std::string itemName = cId;
-        WorldDatabase.escape_string(itemName);
-        QueryResult* result = WorldDatabase.PQuery("SELECT `entry` FROM `item_template` WHERE `name` = '%s'", itemName.c_str());
-        if (!result)
-        {
-            PSendSysMessage(LANG_COMMAND_COULDNOTFIND, cId);
-            SetSentErrorMessage(true);
-            return false;
-        }
-        itemId = result->Fetch()->GetUInt16();
-        delete result;
-    }
-
-    int32 count;
-    if (!ExtractOptInt32(&args, count, 1))
-    {
-        return false;
-    }
-
-    Player* pl = m_session->GetPlayer();
-    Player* plTarget = getSelectedPlayer();
-    if (!plTarget)
-    {
-        plTarget = pl;
-    }
-
-    DETAIL_LOG(GetMangosString(LANG_ADDITEM), itemId, count);
-
-    ItemPrototype const* pProto = ObjectMgr::GetItemPrototype(itemId);
-    if (!pProto)
-    {
-        PSendSysMessage(LANG_COMMAND_ITEMIDINVALID, itemId);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // Subtract
-    if (count < 0)
-    {
-        plTarget->DestroyItemCount(itemId, -count, true, false);
-        PSendSysMessage(LANG_REMOVEITEM, itemId, -count, GetNameLink(plTarget).c_str());
-        return true;
-    }
-
-    // Adding items
-    uint32 noSpaceForCount = 0;
-
-    // check space and find places
-    ItemPosCountVec dest;
-    uint8 msg = plTarget->CanStoreNewItem(NULL_BAG, NULL_SLOT, dest, itemId, count, &noSpaceForCount);
-    if (msg != EQUIP_ERR_OK)                                // convert to possible store amount
-    {
-        count -= noSpaceForCount;
-    }
-
-    if (count == 0 || dest.empty())                         // can't add any
-    {
-        PSendSysMessage(LANG_ITEM_CANNOT_CREATE, itemId, noSpaceForCount);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Item* item = plTarget->StoreNewItem(dest, itemId, true, Item::GenerateItemRandomPropertyId(itemId));
-
-    // Remove binding (let GM give it to another player later)
-    if (pl == plTarget)
-    {
-        for (ItemPosCountVec::const_iterator itr = dest.begin(); itr != dest.end(); ++itr)
-        {
-            if (Item* item1 = pl->GetItemByPos(itr->pos))
-            {
-                item1->SetBinding(false);
-            }
-        }
-    }
-
-    if (count > 0 && item)
-    {
-        pl->SendNewItem(item, count, false, true);
-        if (pl != plTarget)
-        {
-            plTarget->SendNewItem(item, count, true, false);
-        }
-    }
-
-    if (noSpaceForCount > 0)
-    {
-        PSendSysMessage(LANG_ITEM_CANNOT_CREATE, itemId, noSpaceForCount);
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleAddItemSetCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleAddItemSetCommand(char* args)
-{
-    uint32 itemsetId;
-    if (!ExtractUint32KeyFromLink(&args, "Hitemset", itemsetId))
-    {
-        return false;
-    }
-
-    // prevent generation all items with itemset field value '0'
-    if (itemsetId == 0)
-    {
-        PSendSysMessage(LANG_NO_ITEMS_FROM_ITEMSET_FOUND, itemsetId);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* pl = m_session->GetPlayer();
-    Player* plTarget = getSelectedPlayer();
-    if (!plTarget)
-    {
-        plTarget = pl;
-    }
-
-    DETAIL_LOG(GetMangosString(LANG_ADDITEMSET), itemsetId);
-
-    bool found = false;
-    for (uint32 id = 0; id < sItemStorage.GetMaxEntry(); ++id)
-    {
-        ItemPrototype const* pProto = sItemStorage.LookupEntry<ItemPrototype>(id);
-        if (!pProto)
-        {
-            continue;
-        }
-
-        if (pProto->ItemSet == itemsetId)
-        {
-            found = true;
-            ItemPosCountVec dest;
-            InventoryResult msg = plTarget->CanStoreNewItem(NULL_BAG, NULL_SLOT, dest, pProto->ItemId, 1);
-            if (msg == EQUIP_ERR_OK)
-            {
-                Item* item = plTarget->StoreNewItem(dest, pProto->ItemId, true);
-
-                // remove binding (let GM give it to another player later)
-                if (pl == plTarget)
-                {
-                    item->SetBinding(false);
-                }
-
-                pl->SendNewItem(item, 1, false, true);
-                if (pl != plTarget)
-                {
-                    plTarget->SendNewItem(item, 1, true, false);
-                }
-            }
-            else
-            {
-                pl->SendEquipError(msg, NULL, NULL, pProto->ItemId);
-                PSendSysMessage(LANG_ITEM_CANNOT_CREATE, pProto->ItemId, 1);
-            }
-        }
-    }
-
-    if (!found)
-    {
-        PSendSysMessage(LANG_NO_ITEMS_FROM_ITEMSET_FOUND, itemsetId);
-
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleMaxSkillCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleMaxSkillCommand(char* /*args*/)
-{
-    Player* SelectedPlayer = getSelectedPlayer();
-    if (!SelectedPlayer)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // each skills that have max skill value dependent from level seted to current level max skill value
-    SelectedPlayer->UpdateSkillsToMaxSkillsForLevel();
-    return true;
-}
-
-/**
- * @brief Handler for HandleSetSkillCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleSetSkillCommand(char* args)
-{
-    Player* target = getSelectedPlayer();
-    if (!target)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // number or [name] Shift-click form |color|Hskill:skill_id|h[name]|h|r
-    char* skill_p = ExtractKeyFromLink(&args, "Hskill");
-    if (!skill_p)
-    {
-        return false;
-    }
-
-    int32 skill;
-    if (!ExtractInt32(&skill_p, skill))
-    {
-        return false;
-    }
-
-    int32 level;
-    if (!ExtractInt32(&args, level))
-    {
-        return false;
-    }
-
-    int32 maxskill;
-    if (!ExtractOptInt32(&args, maxskill, target->GetPureMaxSkillValue(skill)))
-    {
-        return false;
-    }
-
-    if (skill <= 0)
-    {
-        PSendSysMessage(LANG_INVALID_SKILL_ID, skill);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    SkillLineEntry const* sl = sSkillLineStore.LookupEntry(skill);
-    if (!sl)
-    {
-        PSendSysMessage(LANG_INVALID_SKILL_ID, skill);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    std::string tNameLink = GetNameLink(target);
-
-    if (!target->GetSkillValue(skill))
-    {
-        PSendSysMessage(LANG_SET_SKILL_ERROR, tNameLink.c_str(), skill, sl->name[GetSessionDbcLocale()]);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (level <= 0 || level > maxskill || maxskill <= 0)
-    {
-        return false;
-    }
-
-    target->SetSkill(skill, level, maxskill);
-    PSendSysMessage(LANG_SET_SKILL, skill, sl->name[GetSessionDbcLocale()], tNameLink.c_str(), level, maxskill);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleCombatStopCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleCombatStopCommand(char* args)
-{
-    Player* target;
-    if (!ExtractPlayerTarget(&args, &target))
-    {
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(target))
-    {
-        return false;
-    }
-
-    target->CombatStop();
-    target->GetHostileRefManager().deleteReferences();
-    return true;
-}
-
-/**
- * @brief Handler for HandleRepairitemsCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleRepairitemsCommand(char* args)
-{
-    Player* target;
-    if (!ExtractPlayerTarget(&args, &target))
-    {
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(target))
-    {
-        return false;
-    }
-
-    // Repair items
-    target->DurabilityRepairAll(false, 0, false);
-
-    PSendSysMessage(LANG_YOU_REPAIR_ITEMS, GetNameLink(target).c_str());
-    if (needReportToTarget(target))
-    {
-        ChatHandler(target).PSendSysMessage(LANG_YOUR_ITEMS_REPAIRED, GetNameLink().c_str());
-    }
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyHPCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyHPCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int32 hp = atoi(args);
-    int32 hpm = atoi(args);
-
-    if (hp <= 0 || hpm <= 0 || hpm < hp)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_HP, GetNameLink(chr).c_str(), hp, hpm);
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_HP_CHANGED, GetNameLink().c_str(), hp, hpm);
-    }
-
-    chr->SetMaxHealth(hpm);
-    chr->SetHealth(hp);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyManaCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyManaCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int32 mana = atoi(args);
-    int32 manam = atoi(args);
-
-    if (mana <= 0 || manam <= 0 || manam < mana)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_MANA, GetNameLink(chr).c_str(), mana, manam);
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_MANA_CHANGED, GetNameLink().c_str(), mana, manam);
-    }
-
-    chr->SetMaxPower(POWER_MANA, manam);
-    chr->SetPower(POWER_MANA, mana);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyEnergyCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyEnergyCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int32 energy = atoi(args) * 10;
-    int32 energym = atoi(args) * 10;
-
-    if (energy <= 0 || energym <= 0 || energym < energy)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (!chr)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_ENERGY, GetNameLink(chr).c_str(), energy / 10, energym / 10);
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_ENERGY_CHANGED, GetNameLink().c_str(), energy / 10, energym / 10);
-    }
-
-    chr->SetMaxPower(POWER_ENERGY, energym);
-    chr->SetPower(POWER_ENERGY, energy);
-
-    DETAIL_LOG(GetMangosString(LANG_CURRENT_ENERGY), chr->GetMaxPower(POWER_ENERGY));
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyRageCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyRageCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int32 rage = atoi(args) * 10;
-    int32 ragem = atoi(args) * 10;
-
-    if (rage <= 0 || ragem <= 0 || ragem < rage)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_RAGE, GetNameLink(chr).c_str(), rage / 10, ragem / 10);
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_RAGE_CHANGED, GetNameLink().c_str(), rage / 10, ragem / 10);
-    }
-
-    chr->SetMaxPower(POWER_RAGE, ragem);
-    chr->SetPower(POWER_RAGE, rage);
-
-    return true;
-}
-
-// Edit Player Runic Power
-bool ChatHandler::HandleModifyRunicPowerCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int32 rune = atoi(args) * 10;
-    int32 runem = atoi(args) * 10;
-
-    if (rune <= 0 || runem <= 0 || runem < rune)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_RUNIC_POWER, GetNameLink(chr).c_str(), rune / 10, runem / 10);
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_RUNIC_POWER_CHANGED, GetNameLink().c_str(), rune / 10, runem / 10);
-    }
-
-    chr->SetMaxPower(POWER_RUNIC_POWER, runem);
-    chr->SetPower(POWER_RUNIC_POWER, rune);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyTalentCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyTalentCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    int tp = atoi(args);
-    if (tp < 0)
-    {
-        return false;
-    }
-
-    Unit* target = getSelectedUnit();
-    if (!target)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        // check online security
-        if (HasLowerSecurity((Player*)target))
-        {
-            return false;
-        }
-
-        ((Player*)target)->SetFreeTalentPoints(tp);
-        ((Player*)target)->SendTalentsInfoData(false);
-        return true;
-    }
-    else if (((Creature*)target)->IsPet())
-    {
-        Unit* owner = target->GetOwner();
-        if (owner && owner->GetTypeId() == TYPEID_PLAYER && ((Pet*)target)->IsPermanentPetFor((Player*)owner))
-        {
-            // check online security
-            if (HasLowerSecurity((Player*)owner))
-            {
-                return false;
-            }
-
-            ((Pet*)target)->SetFreeTalentPoints(tp);
-            ((Player*)owner)->SendTalentsInfoData(true);
-            return true;
-        }
-    }
-
-    SendSysMessage(LANG_NO_CHAR_SELECTED);
-    SetSentErrorMessage(true);
-    return false;
-}
-
-/**
- * @brief Handler for HandleModifyASpeedCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyASpeedCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float modSpeed = (float)atof(args);
-
-    if (modSpeed > sWorld.getConfig(CONFIG_UINT32_GM_MAX_SPEED_FACTOR) || modSpeed < 0.1)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    std::string chrNameLink = GetNameLink(chr);
-
-    if (chr->IsTaxiFlying())
-    {
-        PSendSysMessage(LANG_CHAR_IN_FLIGHT, chrNameLink.c_str());
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_ASPEED, modSpeed, chrNameLink.c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_ASPEED_CHANGED, GetNameLink().c_str(), modSpeed);
-    }
-
-    chr->UpdateSpeed(MOVE_WALK, true, modSpeed);
-    chr->UpdateSpeed(MOVE_RUN, true, modSpeed);
-    chr->UpdateSpeed(MOVE_SWIM, true, modSpeed);
-    // chr->UpdateSpeed(MOVE_TURN,   true, modSpeed);
-    chr->UpdateSpeed(MOVE_FLIGHT, true, modSpeed);
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifySpeedCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifySpeedCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float modSpeed = (float)atof(args);
-
-    if (modSpeed > sWorld.getConfig(CONFIG_UINT32_GM_MAX_SPEED_FACTOR) || modSpeed < 0.1)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    std::string chrNameLink = GetNameLink(chr);
-
-    if (chr->IsTaxiFlying())
-    {
-        PSendSysMessage(LANG_CHAR_IN_FLIGHT, chrNameLink.c_str());
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_SPEED, modSpeed, chrNameLink.c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_SPEED_CHANGED, GetNameLink().c_str(), modSpeed);
-    }
-
-    chr->UpdateSpeed(MOVE_RUN, true, modSpeed);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifySwimCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifySwimCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float modSpeed = (float)atof(args);
-
-    if (modSpeed > sWorld.getConfig(CONFIG_UINT32_GM_MAX_SPEED_FACTOR) || modSpeed < 0.01f)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    std::string chrNameLink = GetNameLink(chr);
-
-    if (chr->IsTaxiFlying())
-    {
-        PSendSysMessage(LANG_CHAR_IN_FLIGHT, chrNameLink.c_str());
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_SWIM_SPEED, modSpeed, chrNameLink.c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_SWIM_SPEED_CHANGED, GetNameLink().c_str(), modSpeed);
-    }
-
-    chr->UpdateSpeed(MOVE_SWIM, true, modSpeed);
-
-    return true;
-}
-
-//Edit Player Fly command
-bool ChatHandler::HandleModifyFlyCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float modSpeed = (float)atof(args);
-
-    if (modSpeed > 10.0f || modSpeed < 0.1f)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_FLY_SPEED, modSpeed, GetNameLink(chr).c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_FLY_SPEED_CHANGED, GetNameLink().c_str(), modSpeed);
-    }
-
-    chr->UpdateSpeed(MOVE_FLIGHT, true, modSpeed);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyBWalkCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyBWalkCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float modSpeed = (float)atof(args);
-
-    if (modSpeed > sWorld.getConfig(CONFIG_UINT32_GM_MAX_SPEED_FACTOR) || modSpeed < 0.1f)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    std::string chrNameLink = GetNameLink(chr);
-
-    if (chr->IsTaxiFlying())
-    {
-        PSendSysMessage(LANG_CHAR_IN_FLIGHT, chrNameLink.c_str());
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_CHANGE_BACK_SPEED, modSpeed, chrNameLink.c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_YOURS_BACK_SPEED_CHANGED, GetNameLink().c_str(), modSpeed);
-    }
-
-    chr->UpdateSpeed(MOVE_RUN_BACK, true, modSpeed);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyScaleCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyScaleCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    float Scale = (float)atof(args);
-    if (Scale > 10.0f || Scale <= 0.0f)
-    {
-        SendSysMessage(LANG_BAD_VALUE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    Unit* target = getSelectedUnit();
-    if (target == NULL)
-    {
-        SendSysMessage(LANG_SELECT_CHAR_OR_CREATURE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (target->GetTypeId() == TYPEID_PLAYER)
-    {
-        // check online security
-        if (HasLowerSecurity((Player*)target))
-        {
-            return false;
-        }
-
-        PSendSysMessage(LANG_YOU_CHANGE_SIZE, Scale, GetNameLink((Player*)target).c_str());
-        if (needReportToTarget((Player*)target))
-        {
-            ChatHandler((Player*)target).PSendSysMessage(LANG_YOURS_SIZE_CHANGED, GetNameLink().c_str(), Scale);
-        }
-    }
-
-    target->SetObjectScale(Scale);
-    target->UpdateModelData();
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyMountCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyMountCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    uint16 mId = 1147;
-    float speed = (float)15;
-    uint32 num = atoi(args);
-    switch (num)
-    {
-        case 1:
-            mId = 14340;
-            break;
-        case 2:
-            mId = 4806;
-            break;
-        case 3:
-            mId = 6471;
-            break;
-        case 4:
-            mId = 12345;
-            break;
-        case 5:
-            mId = 6472;
-            break;
-        case 6:
-            mId = 6473;
-            break;
-        case 7:
-            mId = 10670;
-            break;
-        case 8:
-            mId = 10719;
-            break;
-        case 9:
-            mId = 10671;
-            break;
-        case 10:
-            mId = 10672;
-            break;
-        case 11:
-            mId = 10720;
-            break;
-        case 12:
-            mId = 14349;
-            break;
-        case 13:
-            mId = 11641;
-            break;
-        case 14:
-            mId = 12244;
-            break;
-        case 15:
-            mId = 12242;
-            break;
-        case 16:
-            mId = 14578;
-            break;
-        case 17:
-            mId = 14579;
-            break;
-        case 18:
-            mId = 14349;
-            break;
-        case 19:
-            mId = 12245;
-            break;
-        case 20:
-            mId = 14335;
-            break;
-        case 21:
-            mId = 207;
-            break;
-        case 22:
-            mId = 2328;
-            break;
-        case 23:
-            mId = 2327;
-            break;
-        case 24:
-            mId = 2326;
-            break;
-        case 25:
-            mId = 14573;
-            break;
-        case 26:
-            mId = 14574;
-            break;
-        case 27:
-            mId = 14575;
-            break;
-        case 28:
-            mId = 604;
-            break;
-        case 29:
-            mId = 1166;
-            break;
-        case 30:
-            mId = 2402;
-            break;
-        case 31:
-            mId = 2410;
-            break;
-        case 32:
-            mId = 2409;
-            break;
-        case 33:
-            mId = 2408;
-            break;
-        case 34:
-            mId = 2405;
-            break;
-        case 35:
-            mId = 14337;
-            break;
-        case 36:
-            mId = 6569;
-            break;
-        case 37:
-            mId = 10661;
-            break;
-        case 38:
-            mId = 10666;
-            break;
-        case 39:
-            mId = 9473;
-            break;
-        case 40:
-            mId = 9476;
-            break;
-        case 41:
-            mId = 9474;
-            break;
-        case 42:
-            mId = 14374;
-            break;
-        case 43:
-            mId = 14376;
-            break;
-        case 44:
-            mId = 14377;
-            break;
-        case 45:
-            mId = 2404;
-            break;
-        case 46:
-            mId = 2784;
-            break;
-        case 47:
-            mId = 2787;
-            break;
-        case 48:
-            mId = 2785;
-            break;
-        case 49:
-            mId = 2736;
-            break;
-        case 50:
-            mId = 2786;
-            break;
-        case 51:
-            mId = 14347;
-            break;
-        case 52:
-            mId = 14346;
-            break;
-        case 53:
-            mId = 14576;
-            break;
-        case 54:
-            mId = 9695;
-            break;
-        case 55:
-            mId = 9991;
-            break;
-        case 56:
-            mId = 6448;
-            break;
-        case 57:
-            mId = 6444;
-            break;
-        case 58:
-            mId = 6080;
-            break;
-        case 59:
-            mId = 6447;
-            break;
-        case 60:
-            mId = 4805;
-            break;
-        case 61:
-            mId = 9714;
-            break;
-        case 62:
-            mId = 6448;
-            break;
-        case 63:
-            mId = 6442;
-            break;
-        case 64:
-            mId = 14632;
-            break;
-        case 65:
-            mId = 14332;
-            break;
-        case 66:
-            mId = 14331;
-            break;
-        case 67:
-            mId = 8469;
-            break;
-        case 68:
-            mId = 2830;
-            break;
-        case 69:
-            mId = 2346;
-            break;
-        default:
-            SendSysMessage(LANG_NO_MOUNT);
-            SetSentErrorMessage(true);
-            return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (!chr)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    PSendSysMessage(LANG_YOU_GIVE_MOUNT, GetNameLink(chr).c_str());
-    if (needReportToTarget(chr))
-    {
-        ChatHandler(chr).PSendSysMessage(LANG_MOUNT_GIVED, GetNameLink().c_str());
-    }
-
-    chr->SetUInt32Value(UNIT_FIELD_FLAGS, UNIT_FLAG_PVP);
-    chr->Mount(mId);
-
-    WorldPacket data(SMSG_FORCE_RUN_SPEED_CHANGE, (8 + 4 + 1 + 4));
-    data << chr->GetPackGUID();
-    data << (uint32)0;
-    data << (uint8)0;                                       // new 2.1.0
-    data << float(speed);
-    chr->SendMessageToSet(&data, true);
-
-    data.Initialize(SMSG_FORCE_SWIM_SPEED_CHANGE, (8 + 4 + 4));
-    data << chr->GetPackGUID();
-    data << (uint32)0;
-    data << float(speed);
-    chr->SendMessageToSet(&data, true);
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyMoneyCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyMoneyCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* chr = getSelectedPlayer();
-    if (chr == NULL)
-    {
-        SendSysMessage(LANG_NO_CHAR_SELECTED);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(chr))
-    {
-        return false;
-    }
-
-    int32 addmoney = atoi(args);
-
-    uint32 moneyuser = chr->GetMoney();
-
-    if (addmoney < 0)
-    {
-        int32 newmoney = int32(moneyuser) + addmoney;
-
-        DETAIL_LOG(GetMangosString(LANG_CURRENT_MONEY), moneyuser, addmoney, newmoney);
-        if (newmoney <= 0)
-        {
-            PSendSysMessage(LANG_YOU_TAKE_ALL_MONEY, GetNameLink(chr).c_str());
-            if (needReportToTarget(chr))
-            {
-                ChatHandler(chr).PSendSysMessage(LANG_YOURS_ALL_MONEY_GONE, GetNameLink().c_str());
-            }
-
-            chr->SetMoney(0);
-        }
-        else
-        {
-            if (newmoney > MAX_MONEY_AMOUNT)
-            {
-                newmoney = MAX_MONEY_AMOUNT;
-            }
-
-            PSendSysMessage(LANG_YOU_TAKE_MONEY, abs(addmoney), GetNameLink(chr).c_str());
-            if (needReportToTarget(chr))
-            {
-                ChatHandler(chr).PSendSysMessage(LANG_YOURS_MONEY_TAKEN, GetNameLink().c_str(), abs(addmoney));
-            }
-            chr->SetMoney(newmoney);
-        }
-    }
-    else
-    {
-        PSendSysMessage(LANG_YOU_GIVE_MONEY, addmoney, GetNameLink(chr).c_str());
-        if (needReportToTarget(chr))
-        {
-            ChatHandler(chr).PSendSysMessage(LANG_YOURS_MONEY_GIVEN, GetNameLink().c_str(), addmoney);
-        }
-
-        if (addmoney >= MAX_MONEY_AMOUNT)
-        {
-            chr->SetMoney(MAX_MONEY_AMOUNT);
-        }
-        else
-        {
-            chr->ModifyMoney(addmoney);
-        }
-    }
-
-    DETAIL_LOG(GetMangosString(LANG_NEW_MONEY), moneyuser, addmoney, chr->GetMoney());
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyDrunkCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyDrunkCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    uint8 drunkValue = (uint8)atoi(args);
-    if (drunkValue > 100)
-    {
-        drunkValue = 100;
-    }
-
-    if (Player* target = getSelectedPlayer())
-    {
-        target->SetDrunkValue(drunkValue);
-    }
-
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyRepCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyRepCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* target = getSelectedPlayer();
-
-    if (!target)
-    {
-        SendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // check online security
-    if (HasLowerSecurity(target))
-    {
-        return false;
-    }
-
-    uint32 factionId;
-    if (!ExtractUint32KeyFromLink(&args, "Hfaction", factionId))
-    {
-        return false;
-    }
-
-    if (!factionId)
-    {
-        return false;
-    }
-
-    int32 amount = 0;
-    if (!ExtractInt32(&args, amount))
-    {
-        char* rankTxt = ExtractLiteralArg(&args);
-        if (!rankTxt)
-        {
-            return false;
-        }
-
-        std::string rankStr = rankTxt;
-        std::wstring wrankStr;
-        if (!Utf8toWStr(rankStr, wrankStr))
-        {
-            return false;
-        }
-        wstrToLower(wrankStr);
-
-        int r = 0;
-        amount = -42000;
-        for (; r < MAX_REPUTATION_RANK; ++r)
-        {
-            std::string rank = GetMangosString(ReputationRankStrIndex[r]);
-            if (rank.empty())
-            {
-                continue;
-            }
-
-            std::wstring wrank;
-            if (!Utf8toWStr(rank, wrank))
-            {
-                continue;
-            }
-
-            wstrToLower(wrank);
-
-            if (wrank.substr(0, wrankStr.size()) == wrankStr)
-            {
-                int32 delta;
-                if (!ExtractOptInt32(&args, delta, 0) || (delta < 0) || (delta > ReputationMgr::PointsInRank[r] - 1))
-                {
-                    PSendSysMessage(LANG_COMMAND_FACTION_DELTA, (ReputationMgr::PointsInRank[r] - 1));
-                    SetSentErrorMessage(true);
-                    return false;
-                }
-                amount += delta;
-                break;
-            }
-            amount += ReputationMgr::PointsInRank[r];
-        }
-        if (r >= MAX_REPUTATION_RANK)
-        {
-            PSendSysMessage(LANG_COMMAND_FACTION_INVPARAM, rankTxt);
-            SetSentErrorMessage(true);
-            return false;
-        }
-    }
-
-    FactionEntry const* factionEntry = sFactionStore.LookupEntry(factionId);
-
-    if (!factionEntry)
-    {
-        PSendSysMessage(LANG_COMMAND_FACTION_UNKNOWN, factionId);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    if (factionEntry->reputationListID < 0)
-    {
-        PSendSysMessage(LANG_COMMAND_FACTION_NOREP_ERROR, factionEntry->name[GetSessionDbcLocale()], factionId);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    target->GetReputationMgr().SetReputation(factionEntry, amount);
-    PSendSysMessage(LANG_COMMAND_MODIFY_REP, factionEntry->name[GetSessionDbcLocale()], factionId,
-                    GetNameLink(target).c_str(), target->GetReputationMgr().GetReputation(factionEntry));
-    return true;
-}
-
-/**
- * @brief Handler for HandleModifyGenderCommand command.
- *
- * @param args Command arguments.
- * @returns True if the command executed successfully, false otherwise.
- */
-bool ChatHandler::HandleModifyGenderCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* player = getSelectedPlayer();
-
-    if (!player)
-    {
-        PSendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    PlayerInfo const* info = sObjectMgr.GetPlayerInfo(player->getRace(), player->getClass());
-    if (!info)
-    {
-        return false;
-    }
-
-    char* gender_str = args;
-    int gender_len = strlen(gender_str);
-
-    Gender gender;
-
-    if (!strncmp(gender_str, "male", gender_len))           // MALE
-    {
-        if (player->getGender() == GENDER_MALE)
-        {
-            return true;
-        }
-
-        gender = GENDER_MALE;
-    }
-    else if (!strncmp(gender_str, "female", gender_len))    // FEMALE
-    {
-        if (player->getGender() == GENDER_FEMALE)
-        {
-            return true;
-        }
-
-        gender = GENDER_FEMALE;
-    }
-    else
-    {
-        SendSysMessage(LANG_MUST_MALE_OR_FEMALE);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    // Set gender
-    player->SetByteValue(UNIT_FIELD_BYTES_0, 2, gender);
-    player->SetUInt16Value(PLAYER_BYTES_3, 0, uint16(gender) | (player->GetDrunkValue() & 0xFFFE));
-
-    // Change display ID
-    player->InitDisplayIds();
-
-    char const* gender_full = gender ? "female" : "male";
-
-    PSendSysMessage(LANG_YOU_CHANGE_GENDER, player->GetName(), gender_full);
-
-    if (needReportToTarget(player))
-    {
-        ChatHandler(player).PSendSysMessage(LANG_YOUR_GENDER_CHANGED, gender_full, GetNameLink().c_str());
-    }
-
-    return true;
-}
-
-bool ChatHandler::HandleModifyArenaCommand(char* args)
-{
-    if (!*args)
-    {
-        return false;
-    }
-
-    Player* target = getSelectedPlayer();
-    if (!target)
-    {
-        SendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    int32 amount = (int32)atoi(args);
-
-    target->ModifyArenaPoints(amount);
-
-    PSendSysMessage(LANG_COMMAND_MODIFY_ARENA, GetNameLink(target).c_str(), target->GetArenaPoints());
-
-    return true;
-}
